Move namespace on DTO
