+++ DataObjects/DTO/Account/AccountLoginDTO.cs
@@ -1,10 +1,6 @@
-using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
 
-namespace _4kTiles_Backend.DataObjects.DTO.Auth
+namespace _4kTiles_Backend.DataObjects.DTO.Account
 {
     public class AccountLoginDTO
     {
