+++ DataObjects/DTO/Account/AccountRegisterDTO.cs
@@ -1,11 +1,6 @@
-using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text.Json.Serialization;
-using System.Threading.Tasks;
 
-namespace _4kTiles_Backend.DataObjects.DTO.Auth
+namespace _4kTiles_Backend.DataObjects.DTO.Account
 {
     public class AccountRegisterDTO
     {
