+++ Mappers/AccountProfile.cs
@@ -1,5 +1,5 @@
 using _4kTiles_Backend.DataObjects.DAO.Account;
-using _4kTiles_Backend.DataObjects.DTO.Auth;
+using _4kTiles_Backend.DataObjects.DTO.Account;
 using _4kTiles_Backend.Entities;
 using _4kTiles_Backend.Helpers;
